Refactoring: common "shared" pkg, dependencies fixed
