+++ ctrl/shared/ctrl-msg.go
@@ -1,0 +1,8 @@
+package shared
+
+// channel result
+
+type ChResult struct {
+	DummyJson string
+	ConnObj WriterAnswer
+}
