--- ctrl/shared/struct/struct.go
@@ -1,10 +1,0 @@
-package shared_struct
-
-import "github.com/monfron/mapago/ctrl/shared/intf"
-
-// channel result
-
-type ChResult struct {
-	DummyJson string
-	ConnObj shared_intf.WriterAnswer
-}
